[HER-12]Follow-up Rename to .js, change onFailure
